refactor: Clean build system and move from uic to real c++ code

+++ src/common/Settings.h
@@ -13,7 +13,7 @@ struct AppSettings {
   QString defaultSaveDir;
   QString lastCSVDir;
   int observationThreshold;
-  float magnitudeThreshold;
+  double magnitudeThreshold;
   bool createTDFFile;
 };
